"fix(routes) API passageiro e motorista funcionando..."

+++ backend/src/main/java/com/borasio_back/backend/repository/PassageiroRepository.java
@@ -3,5 +3,5 @@ package com.borasio_back.backend.repository;
 import com.borasio_back.backend.model.entity.Passageiro;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PassageiroRepository extends JpaRepository<Passageiro, Integer> {
+public interface PassageiroRepository extends JpaRepository<Passageiro, Long> {
 }
